minor refactor of Event+CoreDataClass #1443
